+++ forms-flow-api/src/formsflow_api/resources/filter.py
@@ -18,12 +18,17 @@ from formsflow_api.services import FilterService
 
 filter_schema = FilterSchema()
 
-API = Namespace("Filter", description="Filter APIs")
+API = Namespace(
+    "Filter",
+    description="Manages filters for review tasks, including creation, retrieval, updating, and deletion.",
+)
 
 criteria = API.model(
     "Criteria",
     {
-        "candidateGroup": fields.String(description="Filter task specific to group"),
+        "candidateGroupsExpression": fields.String(
+            description="Filter task specific to group"
+        ),
         "includeAssignedTasks": fields.Boolean(description="Include assigned task"),
     },
 )
@@ -35,14 +40,24 @@ variable = API.model(
         "label": fields.String(description="Display name"),
     },
 )
-
+task_visible_attributes = API.model(
+    "taskVisibleAttributes",
+    {
+        "applicationId": fields.Boolean(),
+        "assignee": fields.Boolean(),
+        "createdDate": fields.Boolean(),
+        "dueDate": fields.Boolean(),
+        "followUp": fields.Boolean(),
+        "priority": fields.Boolean(),
+    },
+)
 properties = API.model(
     "Properties",
     {"showUndefinedVariable": fields.Boolean(description="Show undefined variables")},
 )
 
-filter_request = API.model(
-    "FilterRequest",
+filter_base_model = API.model(
+    "FilterBaseModel",
     {
         "name": fields.String(description="Name of the filter"),
         "description": fields.String(description="Description about filter"),
@@ -57,13 +72,25 @@ filter_request = API.model(
         "users": fields.List(
             fields.String(), description="Authorized Users to the filter"
         ),
+        "taskVisibleAttributes": fields.Nested(
+            task_visible_attributes, description="Visible attributes in task"
+        ),
+        "order": fields.Integer(description="Filter display order"),
         "parentFilterId": fields.Integer(description="Parent filter id"),
         "filterType": fields.String(description="Filter type"),
     },
 )
+filter_request = API.inherit(
+    "FilterRequest",
+    filter_base_model,
+    {
+        "isMyTasksEnabled": fields.Boolean(),
+        "isTasksForCurrentUserGroupsEnabled": fields.Boolean(),
+    },
+)
 filter_response = API.inherit(
     "FilterResponse",
-    filter_request,
+    filter_base_model,
     {
         "status": fields.String(description="Status of the filter"),
         "tenant": fields.String(description="Authorized Tenant to the filter"),
@@ -75,8 +102,16 @@ filter_response = API.inherit(
     },
 )
 
+filter_response_with_attribute_filters = API.inherit(
+     "FilterResponseWithAttributeFilter",
+     filter_response,
+     {
+         "attributeFilters": fields.List(fields.Nested(filter_response)),
+     }
+)
+
 filter_response_with_default_filter = API.model(
-    "FilterResponse",
+    "FilterResponseWithDefaultFilter",
     {
         "filters": fields.List(fields.Nested(filter_response)),
         "defaultFilter": fields.String(description="Default filter"),
@@ -95,6 +130,7 @@ class FilterResource(Resource):
     @API.doc(
         responses={
             200: "OK:- Successful request.",
+            401: "UNAUTHORIZED:- Authorization header not provided or an invalid token passed.",
             403: "FORBIDDEN:- Permission denied",
         },
         model=[filter_response],
@@ -103,7 +139,7 @@ class FilterResource(Resource):
         """
         Get all filters.
 
-        List all active filters for requests with ```reviewer permission```.
+        List all active filters for requests with ```view filter permission```.
         """
         response, status = FilterService.get_all_filters(), HTTPStatus.OK
         return response, status
@@ -113,19 +149,18 @@ class FilterResource(Resource):
     @profiletime
     @API.doc(
         responses={
-            201: "CREATED:- Successful request.",
+            201: ("CREATED:- Successful request.", filter_response),
             400: "BAD_REQUEST:- Invalid request.",
             401: "UNAUTHORIZED:- Authorization header not provided or an invalid token passed.",
             403: "FORBIDDEN:- Permission denied",
-        },
-        model=filter_response,
+        }
     )
     @API.expect(filter_request)
     def post():
         """
         Create filter.
 
-        Post a new filter using request body for requests with ```reviewer permission```.
+        Post a new filter using request body for requests with ```create filter permission```.
         e.g payload,
         ```
         {
@@ -146,6 +181,17 @@ class FilterResource(Resource):
             },
             "users": [],
             "roles": ["/formsflow/formsflow-reviewer"],
+            "isTasksForCurrentUserGroupsEnabled":true,
+            "isMyTasksEnabled":true,
+            "order": 1,
+            "taskVisibleAttributes": {
+                "applicationId": true,
+                "assignee": true,
+                "createdDate": true,
+                "dueDate": true,
+                "followUp": true,
+                "priority": true
+            }
             "parentFilterId": null,
             "filterType": "TASK"
         }
@@ -170,6 +216,7 @@ class UsersFilterList(Resource):
     @API.doc(
         responses={
             200: "OK:- Successful request.",
+            401: "UNAUTHORIZED:- Authorization header not provided or an invalid token passed.",
             403: "FORBIDDEN:- Permission denied",
         },
         model=filter_response_with_default_filter,
@@ -178,7 +225,7 @@ class UsersFilterList(Resource):
         """
         List filters of current user.
 
-        Get all active filters of current reviewer user for requests with ```reviewer permission```.
+        Get all active filters of current reviewer user for requests with ```view filter permission```.
         """
         response, status = FilterService.get_user_filters(), HTTPStatus.OK
         return response, status
@@ -197,15 +244,16 @@ class FilterResourceById(Resource):
         responses={
             200: "OK:- Successful request.",
             400: "BAD_REQUEST:- Invalid request.",
+            401: "UNAUTHORIZED:- Authorization header not provided or an invalid token passed.",
             403: "FORBIDDEN:- Permission denied",
         },
-        model=filter_response,
+        model=filter_response_with_attribute_filters,
     )
     def get(filter_id: int):
         """
         Get filter by id.
 
-        Get filter details corresponding to a filter id for requests with ```REVIEWER_GROUP``` permission.
+        Get filter details corresponding to a filter id for requests with ```manage all filters``` permission.
         """
         filter_result = FilterService.get_filter_by_id(filter_id)
         response, status = filter_result, HTTPStatus.OK
@@ -219,6 +267,7 @@ class FilterResourceById(Resource):
         responses={
             200: "OK:- Successful request.",
             400: "BAD_REQUEST:- Invalid request.",
+            401: "UNAUTHORIZED:- Authorization header not provided or an invalid token passed.",
             403: "FORBIDDEN:- Permission denied",
         },
         model=filter_response,
@@ -228,7 +277,7 @@ class FilterResourceById(Resource):
         """
         Update filter by id.
 
-        Update filter details corresponding to a filter id for requests with ```REVIEWER_GROUP``` permission.
+        Update filter details corresponding to a filter id for requests with ```create filter``` permission.
         """
         filter_data = filter_schema.load(request.get_json())
         filter_result = FilterService.update_filter(filter_id, filter_data)
@@ -245,6 +294,7 @@ class FilterResourceById(Resource):
         responses={
             200: "OK:- Successful request.",
             400: "BAD_REQUEST:- Invalid request.",
+            401: "UNAUTHORIZED:- Authorization header not provided or an invalid token passed.",
             403: "FORBIDDEN:- Permission denied",
         }
     )
@@ -252,7 +302,7 @@ class FilterResourceById(Resource):
         """
         Delete filter by id.
 
-        Delete filter corresponding to a filter id for requests with ```REVIEWER_GROUP``` permission.
+        Delete filter corresponding to a filter id for requests with ```create filter``` permission.
         """
         FilterService.mark_inactive(filter_id=filter_id)
         response, status = "Deleted", HTTPStatus.OK