+++ forms-flow-data-analysis-api/src/api/resources/checkpoint.py
@@ -5,7 +5,7 @@ from flask_restx import Namespace, Resource
 
 from formsflow_api_utils.utils import cors_preflight
 
-API = Namespace("Checkpoint", description="Checkpoint")
+API = Namespace("Checkpoint", description="Verify the application's health status.")
 
 
 @cors_preflight("GET")