+++ forms-flow-data-analysis-api/src/api/resources/sentiment_analysis.py
@@ -3,7 +3,7 @@ import os
 from http import HTTPStatus
 
 from flask import current_app, jsonify, request
-from flask_restx import Namespace, Resource, cors
+from flask_restx import Namespace, Resource, cors, fields
 
 from formsflow_api_utils.utils import Service, auth, cors_preflight
 
@@ -11,10 +11,45 @@ from api import config
 from api.services.store_sentiment_result import save_sentiment_result
 from api.services.transformers import sentiment_analysis_pipeline_transformers
 
-API = Namespace("sentiment", description="API endpoint for sentiment analysis")
+API = Namespace("SentimentAnalysis", description="API endpoint for sentiment analysis.")
 
 APP_CONFIG = config.get_named_config(os.getenv("DEPLOYMENT_ENV", "production"))
 
+
+sentiment_data_model = API.model(
+    "SentimentData",
+    {
+        "text": fields.String(description="Input Text for sentiment analysis."),
+        "elementId": fields.String(),
+    },
+)
+
+request_model = API.model(
+    "RequestModel",
+    {
+        "applicationId": fields.String(),
+        "formUrl": fields.String(),
+        "data": fields.List(fields.Nested(sentiment_data_model)),
+    },
+)
+
+response_sentiment_data_model = API.inherit(
+    "ResponseSentimentData",
+    sentiment_data_model,
+    {
+        "formUrl": fields.String(),
+        "overallSentiment": fields.String(description="Overall sentiment of the input text."),
+    },
+)
+
+response_model = API.model(
+    "ResponseModel",
+    {
+        "applicationId": fields.String(),
+        "formUrl": fields.String(),
+        "data": fields.List(fields.Nested(response_sentiment_data_model)),
+    },
+)
 
 @cors_preflight("POST,OPTIONS")
 @API.route("", methods=["POST", "OPTIONS"])
@@ -24,8 +59,17 @@ class SentimentAnalysisTransformerResource(Resource):
     @staticmethod
     @cors.crossdomain(origin="*", headers=["Content-Type", "Authorization"])
     @auth.require
+    @API.doc(
+        responses={
+            200: "OK:- Successful request.",
+            400: "BAD_REQUEST:- Invalid request.",
+            401: "UNAUTHORIZED:- Authorization header not provided or an invalid token passed.",
+        },
+        model=response_model
+    )
+    @API.expect(request_model)
     def post():
-        """POST API definition for sentiment analysis API."""
+        """Returns the sentiment (positive, negative, or neutral) for the given text in the request body."""
         try:
             input_json = request.get_json()
             response_json = {